refactor: 없는 페이지와 없는 api 요청 예외 처리 추가 - index로 redirect
