+++ src/main/java/webserver/RequestHandler.java
@@ -7,6 +7,7 @@ import http.request.HttpRequest;
 import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import utils.FileIoUtils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -46,6 +47,14 @@ public class RequestHandler implements Runnable {
             Controller controller = ControllerMapper.map(httpRequest);
             controller.service(httpRequest, httpResponse);
             return;
+        }
+        forward(httpRequest, httpResponse);
+    }
+
+    void forward(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
+        byte[] body = FileIoUtils.loadFileFromClasspath(httpRequest.getPath());
+        if (body == null) {
+            httpResponse.sendRedirect("index.html");
         }
         httpResponse.forward(httpRequest.getPath());
     }
