+++ src/test/java/utils/FileIoUtilsTest.java
@@ -9,7 +9,7 @@ public class FileIoUtilsTest {
 
     @Test
     void loadFileFromClasspath() throws Exception {
-        byte[] body = FileIoUtils.loadFileFromClasspath("./templates/index.html");
+        byte[] body = FileIoUtils.loadFileFromClasspath("/index.html");
         log.debug("file : {}", new String(body));
     }
 }
